# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "S4 RCWA Simulator API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/simulate": {
      "post": {
        "summary": "Simulate",
        "operationId": "simulate_api_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/simulate/orders": {
      "post": {
        "summary": "Simulate Orders",
        "description": "反射の回折次数ごとの角度分布を返す。平面多層膜では 0 次（正反射）のみ。",
        "operationId": "simulate_orders_api_simulate_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/simulate/sweep": {
      "post": {
        "summary": "Simulate Sweep",
        "description": "入射角スイープ。角度ごとの R/T スペクトルと(任意で)反射色を返す。",
        "operationId": "simulate_sweep_api_simulate_sweep_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SweepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/materials": {
      "get": {
        "summary": "List Materials",
        "operationId": "list_materials_api_materials_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaterialSummaryDTO"
                  },
                  "type": "array",
                  "title": "Response List Materials Api Materials Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Material",
        "operationId": "create_material_api_materials_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/materials/{material_id}": {
      "get": {
        "summary": "Get Material",
        "operationId": "get_material_api_materials__material_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Material Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Rename Material",
        "operationId": "rename_material_api_materials__material_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Material Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialSummaryDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Material",
        "operationId": "delete_material_api_materials__material_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Material Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Api Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AngularDistributionDTO": {
        "properties": {
          "wavelengthNm": {
            "type": "number",
            "title": "Wavelengthnm"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/DiffractionOrderDTO"
            },
            "type": "array",
            "title": "Orders"
          }
        },
        "type": "object",
        "required": [
          "wavelengthNm",
          "orders"
        ],
        "title": "AngularDistributionDTO",
        "description": "1 波長の角度分布。orders は出射角の昇順（伝搬する次数のみ）。"
      },
      "ColorDTO": {
        "properties": {
          "r": {
            "type": "integer",
            "title": "R"
          },
          "g": {
            "type": "integer",
            "title": "G"
          },
          "b": {
            "type": "integer",
            "title": "B"
          },
          "hex": {
            "type": "string",
            "title": "Hex"
          }
        },
        "type": "object",
        "required": [
          "r",
          "g",
          "b",
          "hex"
        ],
        "title": "ColorDTO"
      },
      "DiffractionModeDTO": {
        "properties": {
          "R": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "R"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ColorDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "R"
        ],
        "title": "DiffractionModeDTO",
        "description": "回折次数で分けた 1 つの見方（0 次以外 or 全次数）のスペクトルと色。"
      },
      "DiffractionModesDTO": {
        "properties": {
          "nonZeroth": {
            "$ref": "#/components/schemas/DiffractionModeDTO"
          },
          "total": {
            "$ref": "#/components/schemas/DiffractionModeDTO"
          }
        },
        "type": "object",
        "required": [
          "nonZeroth",
          "total"
        ],
        "title": "DiffractionModesDTO",
        "description": "反射率を回折次数で分けたもの。0 次（正反射）は親の R / color。\n\nnon_zeroth は m ≠ 0 の合計（正反射以外の方向へ散った回折光）、total は\n全次数の合計。別方向へ出る光の足し算なので R = total − non_zeroth が成り立つ。\n平面多層膜には m ≠ 0 が無いので non_zeroth は 0、total は R に一致する。"
      },
      "DiffractionOrderDTO": {
        "properties": {
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "angleDeg": {
            "type": "number",
            "title": "Angledeg"
          },
          "reflectance": {
            "type": "number",
            "title": "Reflectance"
          }
        },
        "type": "object",
        "required": [
          "order",
          "angleDeg",
          "reflectance"
        ],
        "title": "DiffractionOrderDTO",
        "description": "1 つの回折次数への反射パワーの配分。"
      },
      "DispersionPointDTO": {
        "properties": {
          "wavelengthNm": {
            "type": "number",
            "title": "Wavelengthnm"
          },
          "n": {
            "type": "number",
            "title": "N"
          },
          "k": {
            "type": "number",
            "title": "K"
          }
        },
        "type": "object",
        "required": [
          "wavelengthNm",
          "n",
          "k"
        ],
        "title": "DispersionPointDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LayerDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "thicknessNm": {
            "type": "number",
            "title": "Thicknessnm"
          },
          "n": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "N"
          },
          "k": {
            "type": "number",
            "title": "K",
            "default": 0.0
          },
          "materialId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materialid"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RegionDTO"
            },
            "type": "array",
            "title": "Regions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "thicknessNm"
        ],
        "title": "LayerDTO",
        "description": "スタック中の 1 層。material_id を指定すると登録した材料を使い、n, k は不要になる。"
      },
      "MaterialCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "content": {
            "type": "string",
            "maxLength": 1048576,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "name",
          "content"
        ],
        "title": "MaterialCreateRequest",
        "description": "材料の登録。content はファイルの中身の文字列（読み取りはバックエンドで行う）。"
      },
      "MaterialDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/DispersionPointDTO"
            },
            "type": "array",
            "title": "Points"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "points",
          "createdAt"
        ],
        "title": "MaterialDTO",
        "description": "詳細用。n, k の表を含む（グラフ表示に使う）。"
      },
      "MaterialRenameRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MaterialRenameRequest"
      },
      "MaterialSummaryDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "MaterialSummaryDTO",
        "description": "一覧用。層の材料選択に必要なのは id と名前だけ。"
      },
      "MeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "MeResponse"
      },
      "OrdersResponse": {
        "properties": {
          "distributions": {
            "items": {
              "$ref": "#/components/schemas/AngularDistributionDTO"
            },
            "type": "array",
            "title": "Distributions"
          }
        },
        "type": "object",
        "required": [
          "distributions"
        ],
        "title": "OrdersResponse"
      },
      "Polarization": {
        "type": "string",
        "enum": [
          "s",
          "p"
        ],
        "title": "Polarization",
        "description": "偏光。S=TE、P=TM。"
      },
      "RegionDTO": {
        "properties": {
          "xNm": {
            "type": "number",
            "title": "Xnm"
          },
          "widthNm": {
            "type": "number",
            "title": "Widthnm"
          },
          "n": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "N"
          },
          "k": {
            "type": "number",
            "title": "K",
            "default": 0.0
          },
          "materialId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materialid"
          }
        },
        "type": "object",
        "required": [
          "xNm",
          "widthNm"
        ],
        "title": "RegionDTO",
        "description": "層内の矩形領域（面内パターン）。単位胞左端を 0 とする nm 座標。\n\nmaterial_id を指定すると登録した材料（波長分散）を使い、n, k は不要になる。"
      },
      "SimulationRequest": {
        "properties": {
          "wlMin": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Wlmin"
          },
          "wlMax": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Wlmax"
          },
          "wlPoints": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Wlpoints"
          },
          "thetaDeg": {
            "type": "number",
            "title": "Thetadeg",
            "default": 0.0
          },
          "pol": {
            "$ref": "#/components/schemas/Polarization",
            "default": "s"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerDTO"
            },
            "type": "array",
            "minItems": 2,
            "title": "Layers"
          },
          "periodNm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodnm"
          },
          "numBasis": {
            "type": "integer",
            "maximum": 61.0,
            "minimum": 1.0,
            "title": "Numbasis",
            "default": 1
          },
          "wavelengthsNm": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "maxItems": 2001,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Wavelengthsnm"
          }
        },
        "type": "object",
        "required": [
          "wlMin",
          "wlMax",
          "wlPoints",
          "layers"
        ],
        "title": "SimulationRequest"
      },
      "SimulationResponse": {
        "properties": {
          "wavelengths": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Wavelengths"
          },
          "R": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "R"
          },
          "T": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "T"
          },
          "reflectedColor": {
            "$ref": "#/components/schemas/ColorDTO"
          },
          "diffraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiffractionModesDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "wavelengths",
          "R",
          "T",
          "reflectedColor"
        ],
        "title": "SimulationResponse"
      },
      "SweepEntryDTO": {
        "properties": {
          "thetaDeg": {
            "type": "number",
            "title": "Thetadeg"
          },
          "R": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "R"
          },
          "T": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "T"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ColorDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "diffraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiffractionModesDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "thetaDeg",
          "R",
          "T"
        ],
        "title": "SweepEntryDTO",
        "description": "1 入射角分の結果。キーは SimulationResponse に合わせて R / T。"
      },
      "SweepRequest": {
        "properties": {
          "wlMin": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Wlmin"
          },
          "wlMax": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Wlmax"
          },
          "wlPoints": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Wlpoints"
          },
          "thetaDeg": {
            "type": "number",
            "title": "Thetadeg",
            "default": 0.0
          },
          "pol": {
            "$ref": "#/components/schemas/Polarization",
            "default": "s"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerDTO"
            },
            "type": "array",
            "minItems": 2,
            "title": "Layers"
          },
          "periodNm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodnm"
          },
          "numBasis": {
            "type": "integer",
            "maximum": 61.0,
            "minimum": 1.0,
            "title": "Numbasis",
            "default": 1
          },
          "wavelengthsNm": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "maxItems": 2001,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Wavelengthsnm"
          },
          "thetaDegs": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 181,
            "minItems": 1,
            "title": "Thetadegs"
          },
          "includeColors": {
            "type": "boolean",
            "title": "Includecolors",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "wlMin",
          "wlMax",
          "wlPoints",
          "layers",
          "thetaDegs"
        ],
        "title": "SweepRequest",
        "description": "入射角スイープのリクエスト。theta_deg の代わりに theta_degs を掃引する。"
      },
      "SweepResponse": {
        "properties": {
          "wavelengths": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Wavelengths"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/SweepEntryDTO"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "wavelengths",
          "entries"
        ],
        "title": "SweepResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}